rename(racing-car): 패키지 및 클래스와 메서드 네이밍 수정
