+++ src/main/java/racingcar/RacingCarMain.java
@@ -3,23 +3,23 @@ package racingcar;
 import java.util.List;
 import racingcar.domain.Car;
 import racingcar.domain.RacingGame;
-import racingcar.ui.RacingCarInput;
-import racingcar.ui.RacingCarOutput;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class RacingCarMain {
     public static void main(String[] args) {
-        List<Car> cars = RacingCarInput.carNameInput();
-        int round = RacingCarInput.roundInput();
+        List<Car> cars = InputView.getCars();
+        int round = InputView.getRound();
         RacingGame racingGame = new RacingGame(cars);
-        RacingCarOutput.printRoundResultMessage();
         raceAndPrintResult(round, racingGame);
-        RacingCarOutput.printWinners(racingGame.findWinners());
+        OutputView.printWinners(racingGame.findWinners());
     }
 
     private static void raceAndPrintResult(int round, RacingGame racingGame) {
+        OutputView.printRoundResultMessage();
         for (int i = 0; i < round; i++) {
             List<Car> raceResult = racingGame.race();
-            RacingCarOutput.printRoundResult(raceResult);
+            OutputView.printRoundResult(raceResult);
         }
     }
 }
